Move RuntimeContextExtensions.LoadModule to Core
'Impl', 'Interop', and 'WinMD' each shipped their own copy of
'RuntimeContextExtensions.LoadModule', with bodies that are
character-for-character identical (only the receiver parameter name
differed). Impl had only the 'PEImage' overload, Interop had only the
'string' overload, WinMD had both.

Consolidate to a single file under 'WinRT.Generator.Core' that carries
both overloads. All three consumer projects pick up the overload they
already used; no behavioral change.

This adds an 'AsmResolver.DotNet' package reference to
'WinRT.Generator.Core' (all 5 generators already reference AsmResolver
directly).

Net diff: 6 files, +63 / -126 LOC (-63 LOC). All 5 generators build
with 0 warnings and pass end-to-end debug-repro validation (ref-gen:
8/8 0 diffs, WinMD: 16-byte MVID-only delta, proj-gen: 327/327
0 diffs, impl-gen: 0 byte diffs, interop-gen: smoke OK).
