+++ src/WinRT.Impl.Generator/Generation/ImplGenerator.cs
@@ -19,6 +19,7 @@ using AsmResolver.PE.DotNet.StrongName;
 using ConsoleAppFramework;
 using WindowsRuntime.Generator;
 using WindowsRuntime.Generator.Errors;
+using WindowsRuntime.Generator.Extensions;
 using WindowsRuntime.Generator.Parsing;
 using WindowsRuntime.ImplGenerator.Errors;
 using WindowsRuntime.ImplGenerator.References;
