+++ src/WinRT.Interop.Generator/Generation/InteropGenerator.Discover.cs
@@ -11,6 +11,7 @@ using AsmResolver.DotNet;
 using AsmResolver.DotNet.Signatures;
 using AsmResolver.PE;
 using WindowsRuntime.Generator.Errors;
+using WindowsRuntime.Generator.Extensions;
 using WindowsRuntime.InteropGenerator.Discovery;
 using WindowsRuntime.InteropGenerator.Errors;
 using WindowsRuntime.InteropGenerator.Models;
